待验证，温度接口统一

+++ HARDWARE/FILTER/filter.c
@@ -2,7 +2,7 @@
  * @Author: huangyouli.scut@gmail.com
  * @Date: 2025-01-11 16:02:24
  * @LastEditors: YouLiHuang huangyouli.scut@gmail.com
- * @LastEditTime: 2025-01-11 16:17:21
+ * @LastEditTime: 2025-01-12 18:02:19
  * @Description:
  *
  * Copyright (c) 2025 by huangyouli, All Rights Reserved.
@@ -11,13 +11,13 @@
 
 /**
  * @description:
- * @param {u16} *input
- * @param {u16} data_len
- * @param {u16} freq_s : sample frequency
- * @param {u16} freq_c ：cut-off frequency
+ * @param {float} *input
+ * @param {float} data_len
+ * @param {float} freq_s : sample frequency
+ * @param {float} freq_c ：cut-off frequency
  * @return {*}
  */
-void low_pass_Filter(u16 *input, u16 data_len, u16 *output, u16 freq_s, u16 freq_c)
+void low_pass_Filter(float *input, float data_len, float *output, float freq_s, float freq_c)
 {
     double alpha = 1.0 / (1.0 + (freq_s / (2.0 * PI * freq_c)));
 
@@ -27,3 +27,4 @@ void low_pass_Filter(u16 *input, u16 data_len, u16 *output, u16 freq_s, u16 freq
         output[i] = alpha * input[i] + (1 - alpha) * output[i - 1];
     }
 }
+
